my new work

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mohamoha <mohamoha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/03 14:51:16 by mohamoha          #+#    #+#             */
-/*   Updated: 2023/07/03 20:31:54 by mohamoha         ###   ########.fr       */
+/*   Updated: 2023/07/06 20:51:29 by mohamoha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
